FIX Don't set session value in endpoint accessed from mail link. (#11686)
With the session cookie's "SameSite" value set to "Strict", we can't set
session values when a request is the result of clicking a link from
another website.
When a user clicks the reset password link from an email within a web
client (e.g. Gmail), this results in not being able to reset their
password.

This change sets the session value _after_ the redirect, while avoiding
putting the reset token into the browser history which could allow a bad
actor to access it.

This change also allows that flow to work for silverstripe/mfa which
needs to alter the second part of the flow, but detects that currently
using the session value's presence or absence.