NEW Allow filtering and excluding based on another list
This provides a convenient shortcut for e.g. filtering a list of pages
by the IDs of a relation, where the relation IDs come from a separate
pre-filtered query.

Using this instead of executing the first query and filtering by IDs
allows the database optimiser to optimise a joint query and reduces
latency overhead of performing separate queries.