FIX Check all classes and don't rely on connection_attempted() (#11807)
DB::is_active() will lazy-connect to the database if we haven't done so
already. That means the DB::connection_attempted() check was actually
preventing us from connecting to the database even though we were ready
to do so.

There's no need to add that check after checking is_active() because
it'll either be active (and we're good to go) or not (and we should say
we're not ready).