ENH Add icon property to Tab (#11885)
For now this is useful for the "More Options" tab in the
save/publish/etc action area in the CMS but should be expanded for other
tabs in the CMS as well.

Note that any front-end usage is the responsibility of project
developers, given that we have no idea how they want to add icons.